feat: use SQLModel and rewrite tests (#63)

+++ iris/agent/prober.py
@@ -3,39 +3,37 @@
 import asyncio
 from multiprocessing import Process
 from pathlib import Path
-from typing import Dict, Optional
-from uuid import UUID
+from typing import Dict
 
 from pycaracal import prober, set_log_level
 
 from iris.agent.settings import AgentSettings
-from iris.commons.redis import AgentRedis
-from iris.commons.schemas.measurements import MeasurementState
+from iris.commons.redis import Redis
 
 
-async def watcher(
-    process: Process, settings: AgentSettings, measurement_uuid: UUID, redis: AgentRedis
+async def watch_cancellation(
+    redis: Redis,
+    process: Process,
+    measurement_uuid: str,
+    agent_uuid: str,
+    interval: float,
 ) -> bool:
-    """Watch the prober execution and stop it according to the measurement state."""
+    """Kill the prober process if the measurement agent is cancelled."""
     while process.is_alive():
-        measurement_state = await redis.get_measurement_state(measurement_uuid)
-        if measurement_state in [
-            MeasurementState.Canceled,
-            MeasurementState.Unknown,
-        ]:
+        if await redis.measurement_agent_cancelled(measurement_uuid, agent_uuid):
             process.kill()
             return False
-        await asyncio.sleep(settings.AGENT_STOPPER_REFRESH)
+        await asyncio.sleep(interval)
     return True
 
 
 def probe(
     settings: AgentSettings,
+    probes_filepath: Path,
     results_filepath: Path,
     round_number: int,
     probing_rate: int,
     prober_statistics: Dict,
-    probes_filepath: Optional[Path] = None,
 ) -> None:
     """Probing interface."""
     # Cap the probing rate if superior to the maximum probing rate
@@ -46,7 +44,7 @@ def probe(
     )
 
     # This set the log level of the C++ logger (spdlog).
-    # This allow the logs to be filtered in C++ (fast)
+    # This allows the logs to be filtered in C++ (fast)
     # before being forwarded to the (slower) Python logger.
     set_log_level(settings.AGENT_CARACAL_LOGGING_LEVEL)
